==> 08_inventor_similarity.ipynb <==
## IMPLEMENTATION OF INVENTOR SIMILARITY
Testing the implementation of obtaining inventor similarity using **node2vec**. <br >
Patent: id: 4136359 Apple Microcomputer

import neo4j 
import pandas as pd
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from patent_neo4j.utils import inventor_to_int

List of Important Patents

df = pd.read_csv("Data/important_patents.csv")
df.head(8)

Set up Neo4j connections and Query Coinventors

conn = Neo4jConnection(uri, user, pwd)
result = conn.query_coinventors(root=df.id[0])

Create file compatible for node2vec

file_name = "./node2vec/graph/" + df.name[0] + ".edgelist"
file_emb = "./node2vec/emb/" + df.name[0] + ".emb"
file_map = "./node2vec/map/" + df.name[0] + ".map"
coinventors = inventor_to_int(result, write=True, file=file_name, mapping=file_map)

Runs node2vec Implementation

import os

command = f'./node2vec/./node2vec -i:{file_name} -o:{file_emb} -l:20 -r:10 -k:10 -e:2 -p:0.1 -d:24'
ret = os.system(command)

if ret != 0:
    print("Error in node2vec!!!")

# Mapping for coinventors
coinventor_mapping = coinventors[1]

Sampling 25 Random Inventors

import random

rand_inventors = random.sample(list(coinventor_mapping.keys()), 25)

### Constructing Dataset for Analysis
This section is to construct a dataset for analysis of using similarity score <br>
- **for each** sampled inventor
    - get **related** inventors
    - obtain the **shortest path** to each inventor
    - **for each** coinventor pair
        - **calculate similarity**

from patent_neo4j.analysis import related_inventors_sp, similarity_score, convert_np

df = pd.DataFrame(columns = ['inventor', 'related', 'hops','sim'])
node_emb = convert_np(file_emb)
for inventor in rand_inventors:
    related_inventors = conn.query_related_inventors(inventor, max_depth=3)
    sp = related_inventors_sp(inventor,related_inventors)
    sim = []
    for index,row in sp.iterrows():
        sim.append(similarity_score(coinventor_mapping, node_emb, row.inventor, row.related))
    sp['sim'] = sim
    df = df.append(sp)

# Saving file to iventor_sim ...
df.dropna(subset=['sim']).to_csv("inventor_sim.csv")
