==> 09_Inventor_Combination.ipynb <==
# INVENTOR COMBINATIONS TO MEASURE INVENTOR SIMILARITY
Measuring the similarity of the inventors, with two distinct types:
1. **Interpatent** Inventor Similarity - Inventor Similarity between patents. Implementation here is between direct ancestor-descendant relationship 
2. **Intrapatent** Inventor Similarity - Similarity between inventors on the patent 

import neo4j 
import pandas as pd
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from patent_neo4j.analysis import get_direct_ancestor
from patent_neo4j.analysis import interpatent_inventor_combination
import itertools

Important Patents List

df = pd.read_csv("Data/important_patents.csv")
df.head(8)

## Inter-Patent Inventor Similarity
**Overall Algorithm**:
1. **Obtain direct ancestor-descendant** relationship
1. **Combine inventors** of the two patents with repeats allowed <br> 
    *("wozniak", "wozniak") would have a similarity score of 1*
1. Compute **average similarity**

Establishing Connection

conn = Neo4jConnection(uri, user, pwd)
result = conn.query_citation_tree(root=df.id[0])

result.head()

# Get Direct Lineage
direct_lineage = get_direct_ancestor(result)

direct_lineage.head()

direct_lineage.shape

direct_lineage = direct_lineage.drop_duplicates()

direct_lineage.shape

Obtain All Inventors from the Built Citation Tree

inventor_tree = conn.query_inventor_tree(root=df.id[0])

inventor_tree.head()

Generate Interpatent Combination

interpatent = interpatent_inventor_combination(direct_ancestor=direct_lineage, inventor_tree=inventor_tree)

interpatent.head()

### Calculation of Similarity
This section assumes the embeddings for the inventors have been persisted in the files below as well as the mapping. <br>
**Important Notes:**
* **node2vec** uses biased random walks, re-running the algorithm would likely yield a different results. However, the results should be consistent
* There exist inventors that **do not have relationships up to 3 hops**. In this case, there would **NOT be any embeddings for such nodes**, resulting a return of None by *similarity_score()*. Such cases are treated as 0

from patent_neo4j.analysis import similarity_score, convert_np
import json

file_emb = "./node2vec/emb/" + df.name[0] + ".emb"
file_map = "./node2vec/map/" + df.name[0] + ".map"
node_emb = convert_np(file_emb)
with open(file_map) as json_file:
    coinventor_mapping = json.load(json_file)

Calculating the **Average Combination Similarity**

inventor_sim = []

for index,row in interpatent.iterrows():
    sim_score = 0
    length = len(row.combination)
    
    for pair in row.combination:
        curr_score = similarity_score(coinventor_mapping, node_emb, pair[0], pair[1])
        if curr_score is not None: # Effectively making no connections to be 0
            sim_score = sim_score + curr_score
        
    if length != 0:
        inventor_sim.append(sim_score/length)
    else:
        inventor_sim.append(0)

interpatent = interpatent.assign(inventor_sim=inventor_sim)

Change Combinations to the Number of Combinations 

interpatent['combination'] = interpatent['combination'].apply(lambda x: len(x))

interpatent.head()

interpatent.to_csv("interpatent.csv")

## Intra-Patent Inventor Similarity

inventor_tree = conn.query_inventor_tree(root=df.id[0])

inventor_tree = inventor_tree.rename({'patent':'id'}, axis=1)

inventor_tree.head()

result = conn.query_citation_tree(root=df.id[0])
direct_lineage = get_direct_ancestor(result)

direct_lineage = direct_lineage.drop_duplicates()

direct_lineage.head()

intrapatent = pd.merge(inventor_tree, direct_lineage, on='id', how='left')

intrapatent['combination'] = intrapatent['inventor'].apply(lambda x: [i for i in itertools.combinations(x, 2)])

intrapatent.head()

intra_sim = []
for index,row in intrapatent.iterrows():
    sim_score = 0
    if len(row.combination) == 0:
        sim_score = 1
    else:
        for p in row.combination:
            sim_score = sim_score + similarity_score(coinventor_mapping, node_emb, p[0], p[1])
        sim_score = sim_score / len(row.combination)
        
    intra_sim.append(sim_score)
intrapatent['intra_sim'] = intra_sim 

intrapatent.head()

intrapatent.drop(columns=['inventor','combination']).to_csv("intrapatent.csv")
