==> 10_JS_div.ipynb <==
# Janson-Shannon Divergence 
Calculating **Janson-Shannon Divergence** (the metric/distance equivalent for Kullback-Leibler Divergence) based on the discrete probability distribution of NBER subcategories. This measure captures the distance between two probability distributions and the main reasoning behind this is as follows <br>
1. Prior work on patents shows limited "cross-category" citation of patents
1. Significant patents are associated with strong generality
1. "Cross-category" citations can be associated with generality of patents
<br>

From this line of reasoning, framing each citation hop as a discrete probability distribution and measuring the divergence allows us to **capture how different the previous citing generation and the next**.

import neo4j 
import pandas as pd
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from patent_neo4j.analysis import assign_missing_nber, nber_distribution, js_divergence
from random import randint

df = pd.read_csv("Data/important_patents.csv")
df.head(8)

conn = Neo4jConnection(uri, user, pwd)
nber = conn.query_nber_category(root=4372077)

nber.shape

len(nber['id'].unique())

nber.head()

## Assigning Missing NBER using "Majority Vote"
Overview of Algorithm:
1. For missing each NBER assignment patent, get all direct ancestors' NBER
1. Based on the counts of NBER categories of direct ancestors, the majority (most votes) category is assigned
1. Repeat until reaches root (which is expected to be non-None)

nber = assign_missing_nber(nber)

nber.head()

## Calculating Distribution
Grouped by hops and NBER categories, the probability for each hops' NBER categories are calculated, both for **main categories** and **subcategores**

distribution = nber_distribution(nber)

## Janson-Shannon Divergence
Overview of Algorithm:
1. For both main and subcategories
1. Generate hop combination
1. Calculate JS Divergence

js = js_divergence(distribution)

for i in js:
    print(i)
