==> 11_Geocoordinates.ipynb <==
# Getting Geo-Coordinates and Calculate Locational Difference
Obtaining the citation trees of a given a patent, the distance (using great circle) is calculated. As for now (08/06/2021), I have no clue what to do with it. But I am interested in characterizing the "dispersion" of location, which is analagous to the dispersion of ideas.

import neo4j 
import pandas as pd
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from patent_neo4j.analysis import geo_distance

df = pd.read_csv("Data/important_patents.csv")
df.head(8)

conn = Neo4jConnection(uri, user, pwd)
geo = conn.query_geo_coordinate(root=df.loc[1,:].id)

### Potential Issue
Dropping NAs is probably **not** random! Most of the missing lat,long are from non-US companies

geo = geo_distance(geo)

geo.head()
