==> 12_Low_Hanging_Fruits.ipynb <==
# Getting Low Hanging Fruits 
Characterizing patents with their citation trees, in this specific part, we look at some easy statistics that can be obtained from the citation trees. As for now, the following attributes of the citation tree is mined:
* Number of Edges (Citing-Cited Relationship)
* Number of Patents
* Edge Density
* Number of Assignees
* Number of Inventors
* Number of Locations
* Average Number of Claims
* Average Similarity of Direct Citing-Cited Relationship

A random sample of 1000 is taken with IDs that range in [4136359,6331415] to control for time (older patents tend to have more citations all else equal). The measures from the important patents list are also included.

import neo4j 
import pandas as pd
import random
from functools import reduce
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection

**Important Patent List**

df = pd.read_csv("Data/important_patents_list.csv", usecols = ["id", "issue_year"])
df.head(5)

**Sampling with Imposed Range**

num_sample = 1000
random_list = random.sample(range(df.loc[0,'id'], df.loc[len(df)-1,'id']), num_sample)

conn = Neo4jConnection(uri, user, pwd)

**Initializing Dataframe**

dataset = pd.DataFrame(columns = ["id","num_edge","num_patents","edge_density",
                                  "unq_assignees","unq_inventors", "unq_loc",
                                  "avg_claims","avg_sim"])

dataset.head()

**Omnibus Function** <br>
All measures are inner functions of this function, with implementation separated for readability

"""
Ugly omnibus function with way too many inner functions.
Used to look worse but I don't want to have too simple functions 
that have no reuse value and the best way to encapsulate it is as such
INPUT:
    citation_tree and patent_id (of root)
OUTPUT:
    List containing [id (root),num_edge, num_patents, edge_density, unq_assignees, unq_inventors,
    unq_loc, avg_claims, avg_sim]
"""
def omnibus_fx(citation_tree, patent_id):
    def count_num_edge(citation_tree):
        # Make Deep Copy of Citation Tree
        num_edges = citation_tree[['id','lineage']].copy()
        # Take direct ancestor (existent of edge)
        num_edges.loc[:,'lineage'] = num_edges.loc[:,'lineage'].apply(lambda x: x[0])
        # Drop duplicates from assignee/inventor information
        num_edges = num_edges.drop_duplicates()
        
        n = len(num_edges)
        
        del num_edges

        # Return |E|
        return n
    
    def count_num_patents(citation_tree):
        patent_set = set(citation_tree['id'])
        num_patent = len(patent_set)
        
        del patent_set

        return num_patent

    def edge_density(num_edges, num_patents):
        d = num_edges/num_patents

        return d
    
    def num_assignees(citation_tree):
        assignee_set = set(citation_tree['assignee'])
        num_assignees = len(assignee_set)
        
        del assignee_set

        return num_assignees
    
    def num_locations(citation_tree):
        location_set = set(citation_tree['location'])
        num_locations = len(location_set)
        
        del location_set

        return num_locations
    
    def num_inventors(citation_tree):
        inventor_set = set(citation_tree['inventor'])
        num_inventors = len(inventor_set)
        
        del inventor_set

        return num_inventors
    
    def avg_similarity(citation_tree):
        director_sim = list(set(citation_tree['similarity'].apply(lambda x: x[0])))
        avg_similarity = sum(director_sim)/len(director_sim)
        
        del director_sim

        return avg_similarity
    
    def avg_claims(citation_tree):
        claims = citation_tree[['id','claims']]
        claims = claims.drop_duplicates()
        claims = claims.dropna()

        num_claims = list(pd.to_numeric(claims['claims']))
        avg_claims = sum(num_claims)/len(num_claims)
        
        del claims

        return avg_claims

    # Using the functions above
    num_edge = count_num_edge(citation_tree)
    num_patents = count_num_patents(citation_tree)
    e_density = edge_density(num_edge, num_patents)
    unq_assignees = num_assignees(citation_tree)
    unq_inventors = num_inventors(citation_tree)
    unq_loc = num_locations(citation_tree)
    avg_claims = avg_claims(citation_tree)
    avg_sim = avg_similarity(citation_tree)
    
    info_list = [patent_id,num_edge,num_patents,e_density,unq_assignees,unq_inventors,unq_loc,avg_claims,avg_sim]
    
    return info_list

Just for a peak

x = "4296981"
citation_tree = conn.query_citation_tree(x)
citation_tree.head()

**Unexpectedly Long Run** <br>
Obtaining Citation Tree and Running Omnibus Function

data = []
for i in range(num_sample):
    citation_tree = conn.query_citation_tree(random_list[i])
    if len(citation_tree) != 0:
        data.append(omnibus_fx(citation_tree,random_list[i]))

dataset = pd.DataFrame(data, columns = ["id","num_edge","num_patents","edge_density",
                                  "unq_assignees","unq_inventors", "unq_loc",
                                  "avg_claims","avg_sim"])

dataset.head()

dataset.to_csv("sample_patents_stats.csv", index=False)

important_patents = list(df['id'])

important_data = []
for i in important_patents:
    citation_tree = conn.query_citation_tree(i)
    print(i)
    important_data.append(omnibus_fx(citation_tree,i))
    del citation_tree

dataset_important = pd.DataFrame(important_data, columns = ["id","num_edge","num_patents","edge_density",
                                  "unq_assignees","unq_inventors", "unq_loc",
                                  "avg_claims","avg_sim"])

dataset_important.to_csv("important_patents_stats.csv", index=False)

dataset_important.head(15)
