==> 15_Generation_Set_Diff.ipynb <==
# Generational Set Difference 
As we study the relationships between the charactersitics of each generation in a citation tree, a question that arose from that is to identify if the number of inventors that we found across generation is reflective of different inventors or are they the same set of inventors. This is applicable for

1. Assignees 
1. Inventors
1. Locations
1. NBER Categories

The measure of similarity for this sense for the first three is the Jaccard-Coefficient, which is measured by $$J(A,B) = \frac{|A \cap B|}{|A \cup B|}$$
where **A** and **B** are sets

import neo4j
import pandas as pd
import random
import numpy as np
import datetime
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from patent_neo4j.analysis import fixing_na_nber, nber_distribution, js_divergence
import math
from patent_neo4j.utils import get_max_generation

### Problematic Patent Roots
In the dataset, there exists many root patents that do not have any citations. These are automatically dropped based on what I have written. There is one root patent that simply is too big for me to process and hence I am just ignoring it as for now

roots = pd.read_csv("./Data/generation.csv")

roots = list(set(roots['root']))

### Getting test and degenerate tree

df = pd.read_csv("Data/Mined Data/sample_patents_stats.csv")

root = df.loc[8, 'id']
degen_root = df.loc[643, 'id']
conn = Neo4jConnection(uri, user, pwd)
citation_tree = conn.query_citation_tree(root)
degen_tree = conn.query_citation_tree(degen_root)

citation_tree.head()

degen_tree.head()

### Fixing Citation Tree
This is equivalent to the section in 14_Generational_Difference 

# Assigning NBER category based on majority vote of parents
citation_tree = fixing_na_nber(citation_tree)
degen_tree = fixing_na_nber(degen_tree)

# Take the max(generation) for a given patent to enforce only single-generation
citation_tree = get_max_generation(citation_tree)
degen_tree = get_max_generation(degen_tree)

citation_tree.head()

degen_tree.head()

### Creating Sets
For assginees, inventors and locations, we are interested to see the sets of assignees, inventors and locations invovled in the citation tree. In this, we aggregate them by generation.

"""
Takes a citation tree, with an optional parameter to identify what to be converted into sets 
and generate sets based on the generation
Input:
    citation_tree and settables
Output:
    generation - dataframe with sets of settables by generation
"""
def setting_settables(citation_tree, settables=['inventor', 'assignee', 'location']):
    
    setter = lambda x: citation_tree.loc[:,['gen',x]].drop_duplicates().groupby('gen').agg({x:lambda y: set(y)}).reset_index()
    
    generation = pd.DataFrame({'gen':[1,2,3]})
    
    for s in settables:
        generation = pd.merge(generation, setter(s), how='left', on='gen')
    
    return generation

set_citation = setting_settables(citation_tree)

set_degen = setting_settables(degen_tree)

### Set Difference Across Generation
For each citation_tree, we compare the sets of inventors, assignees and location using Jaccard Coefficient to observe the difference of each sets. <br>
We look at the combination of generation, which in this case is going to only be
1. Generation 1 and 2 (labeled with 1)
1. Generation 2 and 3 (labeled with 2)
1. Generation 1 and 3 (labeled with 3)

def jaccard_similarity(set_A, set_B):
    
    # Check for NaN conditions
    if type(set_A) is float or type(set_B) is float:
        return 0
    
    # Cardinality of intersection
    n_cap = len(set_A.intersection(set_B))
    
    # Cardinality of union
    n_cup = len(set_A.union(set_B))
    
    # Jaccard Similarity
    return(n_cap/n_cup)

def set_differences(root, set_citation, combination=[(0,1),(1,2),(0,2)], settables=['inventor','assignee','location']):
    
    # Empty list of data
    data = []
    
    # For each combination, return the jaccard similarity of 
    for c in combination:
        gen_rows = [root,c]
        for s in settables:
            gen_rows.append(jaccard_similarity(set_citation.loc[c[0],s], set_citation.loc[c[1],s]))
        data.append(gen_rows)
        
    return data

set_one = set_differences(root,set_citation)
set_two = set_differences(degen_root,set_degen)

data = []
data = data + set_one + set_two
print(data)

df_setdiff = pd.DataFrame(data, columns=['root', 'combination', 'inventor', 'assignee', 'location'])

df_setdiff.head(6)

### Sad Awful Loops

data = []

for root in roots[:500]:
    # Get Citation Tree
    citation_tree = conn.query_citation_tree(root)
    
    # Fixing Citation Tree
    citation_tree = fixing_na_nber(citation_tree)
    citation_tree = get_max_generation(citation_tree)
    
    # Build list
    data = data + set_differences(root,set_citation)
    
df_setdiff = pd.DataFrame(data, columns=['root', 'combination', 'inventor', 'assignee', 'location'])

df_setdiff.head()

df_setdiff.to_csv("gen_set_diff_1.csv", index=False)

data = []

for root in roots[500:]:
    # Get Citation Tree
    citation_tree = conn.query_citation_tree(root)
    
    # Fixing Citation Tree
    citation_tree = fixing_na_nber(citation_tree)
    citation_tree = get_max_generation(citation_tree)
    
    # Build list
    data = data + set_differences(root,set_citation)
    
df_setdiff = pd.DataFrame(data, columns=['root', 'combination', 'inventor', 'assignee', 'location'])

df_setdiff.head()

df_setdiff.to_csv("gen_set_diff_2.csv", index=False)

### Combining the Datasets

df1 = pd.read_csv("gen_set_diff_1.csv")
df2 = pd.read_csv("gen_set_diff_2.csv")

df = df1.append(df2)

df.to_csv("gen_set_diff.csv", index=False)
