==> 16_Inventor_Profiles.ipynb <==
# Inventor Profiles
In this notebook, the we want to look at the behaviours of citation etc. from an inventor perspective. Doing so, we want to know and understand the nature of inventors' profiles for extremely successful patents look like. This notebook is divided to the following sections:
1. Getting inventors of successful profiles
1. Get all patents for the successful inventors
1. Create portfolio **sequential** patents and their behaviours
1. Create portfolio stats

import neo4j
import pandas as pd
import random
import numpy as np
import datetime
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from functools import reduce

## Get Inventors of Successful Patents
Using the important patents dataset, we obtain the set of successful inventors

df = pd.read_csv("./Data/important_patents.csv")

df.head()

roots = df['id']

conn = Neo4jConnection(uri, user, pwd)
patent_info = pd.DataFrame(columns=['id', 'date', 'country', 'claims', 'kind', 'county_fips', 'city','state', 'organization', 'org_type', 'inventor_id', 'attribution_stat'])

for root in roots:
    patent_info = pd.concat([patent_info, conn.query_patent_info(root)])

inventor_set = set(patent_info['inventor_id'])

### Randomly Selecting for Inventors with 10 to 30 Patents

df = pd.read_csv("./Data/ten_to_thirty_patents.csv")

df = df.sample(1000)

inventor_set = set(df['inventor'])

### Top 1 percent Number of Citations 

df = pd.read_csv("./Data/top_one_percent_citations.csv")
conn = Neo4jConnection(uri, user, pwd)
patent_info = pd.DataFrame(columns=['id', 'date', 'country', 'claims', 'kind', 'county_fips', 'city','state', 'organization', 'org_type', 'inventor_id', 'attribution_stat'])

df.head()

roots = df['patent_id']
for root in roots:
    patent_info = pd.concat([patent_info, conn.query_patent_info(root)])

inventor_set = set(patent_info['inventor_id'])

len(inventor_set)

## Patents of Successful Inventors
In this section, we query all of the patents created by the successful inventors. This would allow us to then construct a **portfolio** for the inventors and study the behaviours of the innovation market.

df_patents = pd.DataFrame(columns = ['inventor_id', 'patent_id', 'patent_date'])

for inventor in inventor_set:
    df_patents = pd.concat([df_patents, conn.query_inventor_profile(inventor)])

df_patents.head()

df_patents.groupby(['inventor_id']).count()

## Sequential Patents
Convert dates of patents into UNIX timestamp to allow for the sorting of the data. 

import datetime

df_patents = df_patents.dropna()

df_patents['patent_date'] = df_patents['patent_date'].astype(str).apply(datetime.datetime.strptime, args=("%Y-%m-%d",))

df_patents['patent_date'] = df_patents['patent_date'].apply(datetime.datetime.timestamp)

df_patents.head()

patent_list = list(set(df_patents['patent_id']))

len(patent_list)

## Patent Statistics
Obtain patent level statistics for each of the patent that we have. This would allow the understanding of how the patents of specific inventors behave 

patent_info = conn.query_batch_patent_info(patent_list)

# patent_info returns a list of patent information 
# this line combines them
df_info = reduce(lambda left,right: pd.merge(left,right,on=['patent_id'],
                                            how='outer'), patent_info)

df_info.info()

# Fill the number of citations that are NA to 0
df_info['num_citations'] = df_info['num_citations'].fillna(0)

df_info = df_info.drop_duplicates()

df_info = df_info.groupby('patent_id').agg(lambda x: set(x)).reset_index()

df_info = df_info.applymap(lambda x: x.pop() if (len(x) == 1) else x)

df_info.info()

df_info.head()

df_info = df_info.rename(columns={"inventor_id":"inventor_set"})

## Putting Things Together
In order to understand the behavior of inventor post-peak, we take the lead of the inventor and observe if the future inventor set is the same as the past

df = pd.merge(df_patents, df_info, on='patent_id', how='left')

df.head()

# Evaluate number of inventors in the team as well as the assignees of the patent
df['team_size'] = df['inventor_set'].apply(lambda x: len(x) if type(x) is set else 1)
df['num_assignees'] = df['organization'].apply(lambda x: len(x) if type(x) is set else 1)

# Sort values of inventor and patent date and create "lead" on columns for organization and inventor set
df = df.sort_values(by=["inventor_id", "patent_date"], ascending=[True, True])
shifted = df[['inventor_id', "organization", "inventor_set"]].groupby("inventor_id").shift(1)
df = df.join(shifted.rename(columns=lambda x: x+"_lead"))

# Evaluate if the current inventor set is the same as the previous inventor set
df["same_prev_team"] = (df["inventor_set"] == df["inventor_set_lead"]).astype(int)
df["same_prev_assignee"] = (df["organization"] == df["organization_lead"]).astype(int)

df.head()

columns = ["inventor_id","patent_id","patent_date","num_citations","team_size","num_assignees","same_prev_assignee","same_prev_team"]
df[columns].to_csv("inventor_profiles.csv", index=False)

df.head()

## Inventors and Assignees Over Time
To further investigate the relationship between inventors and assignees, we want to see the *cumulative* inventors and assignees over time.

We introduce a variable called **cumulative inventor** which is the set of all inventors the inventor has worked with up to that point.

# Make sure there is no NA and that all of them are sets
df['inventor_set'] = df['inventor_set'].fillna('0')
df['inventor_set'] = df['inventor_set'].apply(lambda x: set([x]) if (type(x) is not set) else x)

# For every inventor
    # For each patent
        # Add individual inventor set
cumulative_inventor = []
for inventor in inventor_set:
    cumulative_set = set()
    patents = df[df['inventor_id'] == inventor]['patent_id']
    for patent in patents:
        cumulative_set = cumulative_set.union(df[(df['inventor_id'] == inventor) & (df['patent_id'] == patent)].inventor_set.tolist()[0])
        cumulative_inventor.append([inventor, patent, cumulative_set])
        

# Create dataframe for the sets and create counts
cumulative_inv = pd.DataFrame(cumulative_inventor, columns = ['inventor_id', 'patent_id', 'cumulative_inventor'])
cumulative_inv['count_cumulative_inv'] = cumulative_inv['cumulative_inventor'].apply(lambda x: len(x))

# Do the same for organization
df['organization'] = df['organization'].fillna('0')
df['organization'] = df['organization'].apply(lambda x: set([x]) if (type(x) is not set) else x)

# For every inventor
    # For each patent
        # Add individual assignee set
cumulative_assignee = []
for inventor in inventor_set:
    cumulative_set = set()
    patents = df[df['inventor_id'] == inventor]['patent_id']
    for patent in patents:
        cumulative_set = cumulative_set.union(df[(df['inventor_id'] == inventor) & (df['patent_id'] == patent)].organization.tolist()[0])
        cumulative_assignee.append([inventor, patent, cumulative_set])

# Create dataframe for the sets and create counts
cumulative_assignee = pd.DataFrame (cumulative_assignee, columns = ['inventor_id', 'patent_id', 'cumulative_assignee'])
cumulative_assignee['count_cumulative_assignee'] = cumulative_assignee['cumulative_assignee'].apply(lambda x: len(x))

cumulative_inv.head()

cumulative_assignee.head()

# Join tables
columns = ['inventor_id', 'patent_id', 'patent_date', 'num_citations', 'organization', 'inventor_set', 'team_size']
df_cumulative = df[columns].merge(cumulative_inv, how='left', on=['inventor_id', 'patent_id'])
df_cumulative = df_cumulative.merge(cumulative_assignee, how='left', on=['inventor_id', 'patent_id'])

df_cumulative.head(5)

def new_kid_in_block(a,b):
    # Get the set difference and return the length
    diff = a.difference(b)
    return(len(diff))    

# Get lead of cumulative inventor and assignee
shifted = df_cumulative[['inventor_id', "cumulative_inventor", "cumulative_assignee"]].groupby("inventor_id").shift(1)
df_cumulative = df_cumulative.join(shifted.rename(columns=lambda x: x+"_lead"))

# Change into set() cumulative change
df_cumulative['cumulative_inventor_lead'] = df_cumulative['cumulative_inventor_lead'].fillna(0)
df_cumulative['cumulative_inventor_lead'] = df_cumulative['cumulative_inventor_lead'].apply(lambda x: set() if x == 0 else x)
df_cumulative['cumulative_assignee_lead'] = df_cumulative['cumulative_assignee_lead'].fillna(0)
df_cumulative['cumulative_assignee_lead'] = df_cumulative['cumulative_assignee_lead'].apply(lambda x: set() if x == 0 else x)

# Get the size set difference between cumulative inventor vs current 
df_cumulative['new_inventor'] = df_cumulative.apply(lambda x: new_kid_in_block(x.inventor_set, x.cumulative_inventor_lead), axis=1)
df_cumulative['new_assignee'] = df_cumulative.apply(lambda x: new_kid_in_block(x.organization, x.cumulative_assignee_lead), axis=1)

# Save data
columns = ["inventor_id","patent_id","patent_date","num_citations","team_size","count_cumulative_inv","count_cumulative_assignee","new_inventor","new_assignee"]
df_cumulative[columns].to_csv("inventor_profiles_cumulative.csv", index=False)
