==> 17_Switchers.ipynb <==
import neo4j
import pandas as pd
import random
import numpy as np
import datetime
from credentials import uri, user, pwd
from patent_neo4j.connection import Neo4jConnection
from functools import reduce

df = pd.read_csv("./Data/assignee_multiple_patents.csv")

Q1 = df['patent_counts'].quantile(0.25)
Q3 = df['patent_counts'].quantile(0.75)

random.seed(10)
df = df[(df["patent_counts"] <= Q3) & (df["patent_counts"] >= Q1)].sample(2500)

df.head()

assignee_list = list(df["assignee_id"])

conn = Neo4jConnection(uri, user, pwd)

ai = conn.query_assignee_patents(assignee_list=assignee_list)

ai.head()

ai = ai.dropna()
ai["nber_category"] = ai["nber"].apply(lambda x: x[0])

ai['patent_date'] = ai['patent_date'].astype(str).apply(datetime.datetime.strptime, args=("%Y-%m-%d",))

ai = ai.sort_values(by=["assignee_id", "patent_date"], ascending=[True, True]).reset_index()

assignee_set = set(ai["assignee_id"])

ai["switch"] = 0

ai.head()

for a in assignee_set:
    indices = ai[ai["assignee_id"] == a].index
    initial_roots = {ai.iloc[indices[0],6]}
    num_switch = 0
    for i in indices[1:]:
        next_patent_root = ai.iloc[i,6]

        if next_patent_root not in initial_roots:
            num_switch = num_switch + 1
            
        ai.iloc[i,7] = num_switch
        initial_roots = initial_roots.union(next_patent_root) 

ai[ai["assignee_id"] == list(assignee_set)[9]]

ai.to_csv("switchers.csv", index=False)

ai[ai["assignee_id"] == '0002c937-3be3-498d-a07e-3b237bb9aed4'].iloc[i,6] = 1

assignee_info = ai.groupby('assignee_id').agg(list).reset_index()

assignee_info["roots"] = assignee_info["nber_category"].apply(set)
assignee_info["fine_roots"] = assignee_info["nber"].apply(set)

assignee_info.tail(10)

assignee_info["num_roots"] = assignee_info["roots"].apply(len)
assignee_info["num_fine_roots"] = assignee_info["fine_roots"].apply(len)
assignee_info["num_patents"] = assignee_info["patent_id"].apply(len)
assignee_info["total_citation"] = assignee_info["num_citation"].apply(sum)
assignee_info["avg_citation"] = assignee_info["total_citation"]/assignee_info["num_patents"]

assignee_info.head(5)

assignee_info[["num_roots", "num_fine_roots", "num_patents", "total_citation", "avg_citation"]].to_csv("roots_and_patents.csv", index=False)
